fix: int to size_t

+++ src/tokenization.hh
@@ -98,6 +98,6 @@ class Tokenizer {
     inline char consume() { return m_src.at(m_index++); }
 
     const std::string m_src;
-    int m_index{0};
-    int m_line_number{1};
+    size_t m_index{0};
+    size_t m_line_number{1};
 };
